requirements.txt
torch>=2.2
torchvision>=0.17
opencv-python>=4.9
numpy>=1.24
scipy>=1.10
tqdm>=4.66
pyyaml>=6.0
einops>=0.7
pandas>=2.0
scikit-image>=0.22
matplotlib>=3.7
